+++ doc/src/release_notes.rst
@@ -64,6 +64,12 @@ To see all issues & pull requests closed by this release see the
      :start-after: Version 4.1.0
      :end-before: Description
 
+* pgr_maxWeightedMatch
+
+  .. include:: pgr_maxWeightedMatch.rst
+     :start-after: Version 4.1.0
+     :end-before: Description
+
 * pgr_planarFaces
 
   .. include:: pgr_planarFaces.rst
@@ -75,6 +81,7 @@ To see all issues & pull requests closed by this release see the
 
 * pgr_coreNumbers
 * pgr_makeBiconnectedPlanar
+* :issue:3136: pgr_maxWeightedMatch
 * pgr_planarFaces
 
 4.1.0 Code enhancements
@@ -97,7 +104,7 @@ To see all issues & pull requests closed by this release see the
   process/driver pair
 * :issue:`3140`: New function: pgr_makeBiconnectedPlannar
 * :issue:`3142`: New function: pgr_coreNumbers
-* :issue:`3142`: New function: pgr_planarFaces
+* :issue:`3143`: New function: pgr_planarFaces
 * :issue:`3154`: betweennessCentrality: use allpairs/process and driver
 
 4.1.0 Bug Fixes